Split functions into own files
